requirements.txt
asgiref==3.3.4
autopep8==1.5.6
dateutils==0.6.12
Django==3.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
Pillow==8.2.0
pycodestyle==2.7.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
